requirements.txt
streamlit
Pillow
IPython.display
matplotlib.pyplot
numpy
io
tensorflow
tensorflow_hub
